renamed android package to com.armandoassuncao
