+++ README.md
@@ -22,7 +22,7 @@
 #### Android
 
 1. Open up `android/app/src/main/java/[...]/MainActivity.java`
-  - Add `import com.reactlibrary.RNCaMasPackage;` to the imports at the top of the file
+  - Add `import com.armandoassuncao.RNCaMasPackage;` to the imports at the top of the file
   - Add `new RNCaMasPackage()` to the list returned by the `getPackages()` method
 2. Append the following lines to `android/settings.gradle`:
   	```
