+++ android/src/main/java/com/armandoassuncao/RNCaMasPackage.java
@@ -1,4 +1,4 @@
-package com.reactlibrary;
+package com.armandoassuncao;
 
 import java.util.Arrays;
 import java.util.Collections;
