+++ examples/CAMASTestApp/android/app/src/main/java/com/camastestapp/MainApplication.java
@@ -9,7 +9,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
-import com.reactlibrary.RNCaMasPackage;
+import com.armandoassuncao.RNCaMasPackage;
 
 import java.util.Arrays;
 import java.util.List;
